python/v2_0_5/requirements.txt
mysql-connector-python==1.0.7

